requirements.txt
pickle-mixin
requests
pycryptodome
fido2
pyscard
pyserial

